improvement: simplify focus and delay on start (#5359)
This removes the delay to show cells when in edit mode (only needed for
run mode).
