+++ frontend/src/components/editor/renderers/vertical-layout/__tests__/useDelayVisibility.test.ts
@@ -2,7 +2,6 @@
 
 import { act, renderHook } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import * as cellsModule from "@/core/cells/cells";
 import { useDelayVisibility } from "../useDelayVisibility";
 
 describe("useDelayVisibility", () => {
@@ -19,100 +18,70 @@ describe("useDelayVisibility", () => {
   });
 
   it("should start with invisible state", () => {
-    const { result } = renderHook(() => useDelayVisibility(5, "edit"));
+    const { result } = renderHook(() => useDelayVisibility(5, "read"));
     expect(result.current.invisible).toBe(true);
   });
 
-  it("should become visible after delay", () => {
-    const { result } = renderHook(() => useDelayVisibility(5, "edit"));
+  it("should become visible after delay in read mode", () => {
+    const { result } = renderHook(() => useDelayVisibility(5, "read"));
 
     expect(result.current.invisible).toBe(true);
 
-    // Advance timers by the calculated delay (5-1)*15 = 60ms
+    // Advance timers by the calculated delay (5-1)*30 = 120ms
     act(() => {
-      vi.advanceTimersByTime(60);
+      vi.advanceTimersByTime(120);
     });
 
     expect(result.current.invisible).toBe(false);
   });
 
-  it("should cap delay at 100ms for large number of cells", () => {
-    const { result } = renderHook(() => useDelayVisibility(20, "edit"));
+  it("should clamp delay at minimum 100ms for small number of cells", () => {
+    const { result } = renderHook(() => useDelayVisibility(2, "read"));
 
     expect(result.current.invisible).toBe(true);
 
-    // Advance timers by 99ms (not enough)
+    // Advance timers by 99ms (not enough, minimum is 100ms)
     act(() => {
       vi.advanceTimersByTime(99);
     });
     expect(result.current.invisible).toBe(true);
 
-    // Advance remaining 1ms to reach 100ms cap
+    // Advance remaining 1ms to reach 100ms minimum
     act(() => {
       vi.advanceTimersByTime(1);
     });
     expect(result.current.invisible).toBe(false);
   });
 
-  it("should not focus any cell in read mode", () => {
-    const focusSpy = vi.spyOn(cellsModule, "getNotebook");
+  it("should clamp delay at maximum 2000ms for large number of cells", () => {
+    const { result } = renderHook(() => useDelayVisibility(100, "read"));
 
-    renderHook(() => useDelayVisibility(5, "read"));
+    expect(result.current.invisible).toBe(true);
 
+    // Advance timers by 1999ms (not enough)
     act(() => {
-      vi.advanceTimersByTime(100);
+      vi.advanceTimersByTime(1999);
     });
+    expect(result.current.invisible).toBe(true);
 
-    expect(focusSpy).not.toHaveBeenCalled();
+    // Advance remaining 1ms to reach 2000ms cap
+    act(() => {
+      vi.advanceTimersByTime(1);
+    });
+    expect(result.current.invisible).toBe(false);
   });
 
-  it("should focus cell from URL if scrollTo parameter exists", () => {
-    // Mock location.hash
-    const originalHash = window.location.hash;
-    Object.defineProperty(window, "location", {
-      value: { hash: "#scrollTo=testCell" },
-      writable: true,
-    });
+  it("should not apply delay in edit mode", () => {
+    const { result } = renderHook(() => useDelayVisibility(5, "edit"));
 
-    // Mock document.querySelector
-    const mockElement = document.createElement("div");
-    mockElement.scrollIntoView = vi.fn();
-    mockElement.focus = vi.fn();
-    mockElement.dataset.cellId = "cell-123";
+    // In edit mode, the delay should not be applied, so it should immediately be visible
+    expect(result.current.invisible).toBe(false);
+  });
 
-    const querySelectorSpy = vi
-      .spyOn(document, "querySelector")
-      .mockReturnValue(mockElement as unknown as HTMLElement);
+  it("should not apply delay in present mode", () => {
+    const { result } = renderHook(() => useDelayVisibility(5, "present"));
 
-    // Mock getNotebook
-    const mockEditor = { focus: vi.fn() };
-    vi.spyOn(cellsModule, "getNotebook").mockReturnValue({
-      cellIds: { iterateTopLevelIds: [] },
-      cellData: {},
-      cellHandles: {
-        "cell-123": { current: { editorView: mockEditor } },
-      },
-    } as unknown as cellsModule.NotebookState);
-
-    renderHook(() => useDelayVisibility(5, "edit"));
-
-    act(() => {
-      vi.advanceTimersByTime(100);
-    });
-
-    expect(querySelectorSpy).toHaveBeenCalledWith(
-      '[data-cell-name="testCell"]',
-    );
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    expect(mockElement.scrollIntoView).toHaveBeenCalled();
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    expect(mockElement.focus).toHaveBeenCalled();
-    expect(mockEditor.focus).toHaveBeenCalled();
-
-    // Restore original hash
-    Object.defineProperty(window, "location", {
-      value: { hash: originalHash },
-      writable: true,
-    });
+    // In present mode, the delay should not be applied
+    expect(result.current.invisible).toBe(false);
   });
 });
