+++ frontend/src/core/run-app.tsx
@@ -4,12 +4,15 @@ import { useAtomValue } from "jotai";
 import { useEffect } from "react";
 import { AppContainer } from "@/components/editor/app-container";
 import { AppHeader } from "@/components/editor/header/app-header";
+import { Spinner } from "@/components/icons/spinner";
+import { DelayMount } from "@/components/utils/delay-mount";
 import { CellsRenderer } from "../components/editor/renderers/cells-renderer";
 import { notebookIsRunningAtom, useCellActions } from "./cells/cells";
 import type { AppConfig } from "./config/config-schema";
 import { RuntimeState } from "./kernel/RuntimeState";
 import { getSessionId } from "./kernel/session";
 import { sendComponentValues } from "./network/requests";
+import { isAppConnecting } from "./websocket/connection-utils";
 import { useMarimoWebSocket } from "./websocket/useMarimoWebSocket";
 
 interface AppProps {
@@ -34,6 +37,23 @@ export const RunApp: React.FC<AppProps> = ({ appConfig }) => {
   });
 
   const isRunning = useAtomValue(notebookIsRunningAtom);
+  const isConnecting = isAppConnecting(connection.state);
+
+  const renderCells = () => {
+    // If we are connecting for more than 2 seconds, show a spinner
+    if (isConnecting) {
+      return (
+        <DelayMount milliseconds={2000} fallback={null}>
+          <Spinner className="mx-auto" />
+          <p className="text-center text-sm text-muted-foreground mt-2">
+            Connecting...
+          </p>
+        </DelayMount>
+      );
+    }
+
+    return <CellsRenderer appConfig={appConfig} mode="read" />;
+  };
 
   return (
     <AppContainer
@@ -42,7 +62,7 @@ export const RunApp: React.FC<AppProps> = ({ appConfig }) => {
       width={appConfig.width}
     >
       <AppHeader connection={connection} className={"sm:pt-8"} />
-      <CellsRenderer appConfig={appConfig} mode="read" />
+      {renderCells()}
     </AppContainer>
   );
 };
